requirements.txt
boto3==1.39.10
botocore==1.39.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
moto==5.1.8
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
responses==0.25.7
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
xmltodict==0.14.2